+++ geotrek/infrastructure/filters.py
@@ -20,6 +20,11 @@ class InfrastructureFilterSet(StructureRelatedFilterSet):
         field = self.form.fields['type']
         field.queryset = field.queryset.exclude(type=INFRASTRUCTURE_TYPES.SIGNAGE)
 
+        field = self.form.fields['type__type']
+        all_choices = field.widget.choices
+        all_choices = [c for c in all_choices if c[0] != INFRASTRUCTURE_TYPES.SIGNAGE]
+        field.widget.choices = [('', _(u"Category"))] + all_choices
+
     class Meta(StructureRelatedFilterSet.Meta):
         model = Infrastructure
         fields = StructureRelatedFilterSet.Meta.fields + ['type__type', 'type']
@@ -29,11 +34,6 @@ class SignageFilterSet(StructureRelatedFilterSet):
     intervention_year = YearFilter(name='interventions_set__date',
                                    widget=InfrastructureYearSelect)
 
-    def __init__(self, *args, **kwargs):
-        super(SignageFilterSet, self).__init__(*args, **kwargs)
-        field = self.form.fields['type']
-        field.queryset = field.queryset.filter(type=INFRASTRUCTURE_TYPES.SIGNAGE)
-
     class Meta(StructureRelatedFilterSet.Meta):
         model = Signage
-        fields = StructureRelatedFilterSet.Meta.fields + ['type']
+        fields = StructureRelatedFilterSet.Meta.fields